requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.27
python-dotenv==1.0.1
faiss-cpu==1.11.0
pydantic==2.11.7
openai==1.93.3 